Implement Github search api for repositories
The purpose of this commit is to extend GitHub API rate to 10 requests per minute by using GitHub search api(the main benefit is that the search api has it's own rate limit). By this way we also will have the advantage to query the organization repositories instead of getting all repositories for organization and filtering them on client.